braze batching
