+++ packages/destination-actions/src/destinations/braze/trackPurchase/index.ts
@@ -1,21 +1,7 @@
-import { omit, IntegrationError } from '@segment/actions-core'
 import type { ActionDefinition } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
-import dayjs from '../../../lib/dayjs'
-import { getUserAlias } from '../userAlias'
-
-type DateInput = string | Date | number | null | undefined
-type DateOutput = string | undefined | null
-
-function toISO8601(date: DateInput): DateOutput {
-  if (date === null || date === undefined) {
-    return date
-  }
-
-  const d = dayjs(date)
-  return d.isValid() ? d.toISOString() : undefined
-}
+import { sendTrackPurchase } from '../utils'
 
 const action: ActionDefinition<Settings, Payload> = {
   title: 'Track Purchase',
@@ -118,49 +104,10 @@ const action: ActionDefinition<Settings, Payload> = {
     }
   },
   perform: (request, { settings, payload }) => {
-    const { braze_id, external_id } = payload
-
-    // Extract valid user_alias shape. Since it is optional (oneOf braze_id, external_id) we need to only include it if fully formed.
-    const user_alias = getUserAlias(payload.user_alias)
-
-    if (!braze_id && !user_alias && !external_id) {
-      throw new IntegrationError(
-        'One of "external_id" or "user_alias" or "braze_id" is required.',
-        'Missing required fields',
-        400
-      )
-    }
-
-    // Skip when there are no products to send to Braze
-    if (payload.products.length === 0) {
-      return
-    }
-
-    const reservedKeys = Object.keys(action.fields.products.properties ?? {})
-    const properties = omit(payload.properties, reservedKeys)
-
-    const base = {
-      braze_id,
-      external_id,
-      user_alias,
-      app_id: settings.app_id,
-      time: toISO8601(payload.time),
-      properties,
-      _update_existing_only: payload._update_existing_only
-    }
-
-    const purchases = payload.products.map((product) => ({
-      ...base,
-      product_id: product.product_id,
-      currency: product.currency ?? 'USD',
-      price: product.price,
-      quantity: product.quantity
-    }))
-
-    return request(`${settings.endpoint}/users/track`, {
-      method: 'post',
-      json: { purchases }
-    })
+    return sendTrackPurchase(request, settings, [payload])
+  },
+  performBatch: (request, { settings, payload }) => {
+    return sendTrackPurchase(request, settings, payload)
   }
 }
